SS-9745: extract createModelStateCore to dedupe hook vs hook-free port
